No Comments

+++ Assets/scripts/Action.cs
@@ -15,6 +15,6 @@ public class Action {
 	public int[] meterCost;
 
     public int gAngle;
-    public int gStrength;
+    public float gStrength;
 
 }
